branding-ai-marketing/requirements.txt
# ===================================================================
# AI 브랜드 컨설팅 서비스 - 의존성 패키지
# ===================================================================
#
# 📌 설치 방법:
#    pip install -r requirements.txt
#
# 🔍 Python 버전 확인:
#    python --version
#    (Python 3.13 이상 필수)
#
# ===================================================================
# Python 버전 요구사항
# ===================================================================
# Python >= 3.13 (GPT-5.1 및 최신 AI 모델 지원)

# ===================================================================
# LangGraph & LangChain - AI 워크플로우 및 상태 관리
# ===================================================================
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5

# ===================================================================
# AI 모델 API
# ===================================================================
openai>=1.10.0              # OpenAI GPT-5.1, DALL-E-3
google-genai>=1.0.0         # Google Gemini 3 Pro (이미지 생성)

# ===================================================================
# 데이터베이스 - SQLAlchemy 및 PostgreSQL
# ===================================================================
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter

# ===================================================================
# FastAPI - REST API 서버
# ===================================================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# ===================================================================
# Pydantic - 데이터 검증 및 직렬화
# ===================================================================
pydantic>=2.0.0

# ===================================================================
# 유틸리티
# ===================================================================
python-dotenv>=1.0.0  # 환경 변수 관리
requests>=2.31.0      # HTTP 요청
cloudinary>=1.36.0    # 이미지 호스팅 및 관리

# ===================================================================
# 개발 도구 (선택사항)
# ===================================================================
pytest>=7.4.0   # 테스트 프레임워크
black>=23.0.0   # 코드 포맷터
